% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs.matrix.R
\name{cs.matrix}
\alias{cs.matrix}
\title{Cosine Similarity Matrix}
\usage{
cs.matrix(
  x,
  words,
  word_embeddings,
  method = "max",
  parallel = FALSE,
  n.cluster = NULL,
  sparse = FALSE
)
}
\arguments{
\item{x}{a tibble containing 2 columns; line and word.  The 'line' column contains the observation number
that the word from the 'word' column appears in.  See 'Examples'.}

\item{words}{a character vector of words that will represent the columns of the resulting matrix.}

\item{word_embeddings}{named list of word embeddings.  See \code{\link{formatWordEmbeddings}}.}

\item{method}{function to apply across each column.  Options include \code{c("max", "sum", "mean")}.}

\item{parallel}{logical, indicating if the matrix should be calculated in parallel.}

\item{n.cluster}{integer, the number of clusters to use if \code{parallel=TRUE}.}

\item{sparse}{logical, indicating if a sparse matrix should be returned.}
}
\value{
a (sparse) matrix of cosine similarities
}
\description{
Creates the design matrix of cosine similarities from textual observations and a
vector of words.
}
\details{
A function to create a design matrix of cosine similarities from textual observations
and a vector of words.  The resulting matrix will be of dimension \eqn{unique(x$line) \times length(words)}.

Consider 2 words with word embedding representations \eqn{a} and \eqn{b}.  Then the cosine similarity
is defined as \deqn{sim_cos(a,b)=\frac{a \cdot b}{|| a ||_2 \cdot || b ||_2}}{sim_cos(a,b)=(a \cdot b)/(|| a ||_2 \cdot || b ||_2)}.

If \code{method = "max"}, for a given line with \eqn{m} words, each row of the returned matrix is defined as \eqn{max_{i=1,...,m}(sim_cos(a_j, b_i))}.
\code{method = "sum"} or \code{method = "mean"} are defined
in a similar fashion.
}
\examples{
\dontrun{
require(dplyr)
require(tidytext)

word_embeddings <- formatWordEmbeddings(embedding_matrix_example, normalize = TRUE)


sentences <- data.frame("Description" = c("Statistics is great!",
                                          "My dog is fluffy.",
                                          "What is your favorite class?"),
                        stringsAsFactors = FALSE)
x <- tibble(line = 1:nrow(sentences), text = sentences$Description) \%>\%
  unnest_tokens(word, text)

cs.matrix(x, words = c("stats", "cat"), word_embeddings)
}

}
\references{
Goldberg, Y. (2017) \emph{Neural Network Methods for Natural Language Processing.} San Rafael, CA: Morgan & Claypool Publishers.
}
\seealso{
\code{\link{cs}}, \code{\link{formatWordEmbeddings}}
}
