% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs.R
\name{cs}
\alias{cs}
\title{Cosine Similarity}
\usage{
cs(a, b, word_embeddings)
}
\arguments{
\item{a, b}{characters or character vectors containing words in \code{word_embeddings}.}

\item{word_embeddings}{named list of word embeddings.  See \code{\link{formatWordEmbeddings}}.}
}
\value{
a matrix of cosine similarities
}
\description{
This function finds the cosine similarity between two vectors of words.
}
\details{
Consider 2 words with word embedding representations \eqn{a} and \eqn{b}.  Then the cosine similarity
is defined as \deqn{sim_cos(a,b)=\frac{a \cdot b}{|| a ||_2 \cdot || b ||_2}}{sim_cos(a,b)=(a \cdot b)/(|| a ||_2 \cdot || b ||_2)}

If \eqn{A = (a_1,...,a_n)} and \eqn{B = (b_1,...,b_m)}, then the result
is a matrix of \eqn{m \times n} dimension with each entry in cell (i, j) defined as \eqn{sim_cos(a_j, b_i)}.
}
\examples{
\dontrun{

a <- "home"
b <- "house"
cs(a, b, word_embeddings)

a <- c("home", "apartment", "mansion")
b <- c("my", "dog", "sleeps", "in", "her", "dog", "house")
cs(a, b, word_embeddings)
}

}
\references{
Goldberg, Y. (2017) \emph{Neural Network Methods for Natural Language Processing.} San Rafael, CA: Morgan & Claypool Publishers.
}
\seealso{
\code{\link{formatWordEmbeddings}}
}
