% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding_matrix_example.R
\docType{data}
\name{embedding_matrix_example}
\alias{embedding_matrix_example}
\title{Embedding Matrix Example}
\format{A \code{data.frame} in the format needed for \code{\link{formatWordEmbeddings}}}
\usage{
data(embedding_matrix_example)
}
\description{
An example word embedding matrix containing words used in examples. The embeddings
are taken from the \code{glove.6B.300d.txt} file from \url{https://nlp.stanford.edu/projects/glove/}.
}
\references{
Jeffrey Pennington, Richard Socher, and Christopher D. Manning. 2014. GloVe: Global Vectors for Word Representation. \url{https://nlp.stanford.edu/projects/glove/}.
}
\keyword{datasets}
