% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatWordEmbeddings.R
\name{formatWordEmbeddings}
\alias{formatWordEmbeddings}
\title{Format Word Embeddings}
\usage{
formatWordEmbeddings(embedding_matrix, normalize = TRUE, verbose = TRUE)
}
\arguments{
\item{embedding_matrix}{word embedding matrix. For a matrix containing information on \eqn{n} words, with
each word being represented by a \eqn{d} dimensional vector, \code{embedding_matrix} should have
\eqn{n} rows and \eqn{d+1} columns where the first column contains the words.}

\item{normalize}{logical; should the word embeddings be normalized.}

\item{verbose}{logical; should the function report on progress.}
}
\value{
A named list of word embeddings.
}
\description{
This function formats the word embeddings.
}
\details{
This function downloads GloVe (https://nlp.stanford.edu/projects/glove/)
and formats the word embeddings.  The result is a named list of word embeddings.  Each
entry in the list is a numeric vector of length \code{dimension} representing the word
embedding for that entry's name (see examples).
}
\examples{
\dontrun{
# temp <- tempfile()
# download.file("http://nlp.stanford.edu/data/wordvecs/glove.6B.zip", temp)

# embedding_matrix <- read.table(unz(temp, "glove.6B.300d.txt"), quote = "",
#                                comment.char = "", stringsAsFactors = FALSE)

word_embeddings <- formatWordEmbeddings(embedding_matrix_example, normalize = TRUE, verbose = TRUE)

# Extract the word embedding for "the"
word_embeddings[["the"]]
}

}
\references{
Jeffrey Pennington, Richard Socher, and Christopher D. Manning. 2014. GloVe: Global Vectors for Word Representation. \url{https://nlp.stanford.edu/projects/glove/}.
}
